pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "outlook-cli"
version = "0.1.0"
description = "Outlook 365 CLI — read, send, and manage emails from the terminal"
readme = "README.md"
license = {text = "MIT"}
authors = [{name = "Yusuf Altunbicak"}]
requires-python = ">=3.10"
keywords = ["outlook", "email", "cli", "office365", "calendar"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Communications :: Email",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "PyYAML>=6.0",
    "httpx>=0.25",
    "playwright>=1.40",
    "beautifulsoup4>=4.12",
]

[project.urls]
Homepage = "https://github.com/yusufaltunbicak/outlook-cli"
Repository = "https://github.com/yusufaltunbicak/outlook-cli"
Issues = "https://github.com/yusufaltunbicak/outlook-cli/issues"

[project.scripts]
outlook = "outlook_cli.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["outlook_cli"]
